+++ lib/kb_gtdbtk/core/genome_obj_update.py
@@ -85,6 +85,7 @@ def check_obj_type_assembly(
 # update_genome_assembly_objs_class()
 #
 def update_genome_assembly_objs_class(
+        primary_wsid: int,
         upa: str,
         classification: Dict[str, str],
         overwrite_tax: int,
@@ -99,7 +100,10 @@ def update_genome_assembly_objs_class(
     KBaseSets.GenomeSet
     KBaseSearch.GenomeSet
     KBaseGenomes.Genome
+    KBaseGenomeAnnotations.Assembly
+    KBaseSets.AssemblySet
 
+    :param primary_wsid: The ID of the calling workspace.  Important for saving objects from other workspaces
     :param upa: The KBase UPA (e.g. X/Y/Z format) for the object from which sequence data will be
         downloaded.
     :param classification: a dict with name key to GTDB classification
@@ -123,7 +127,7 @@ def update_genome_assembly_objs_class(
         else:
             raise ValueError(f'{obj_type} type is not supported')
 
-        return process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients)
+        return process_genome_objs(primary_wsid, top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients)
 
     elif check_obj_type_assembly (obj_type):
         if 'KBaseSets.AssemblySet' == obj_type:
@@ -133,7 +137,7 @@ def update_genome_assembly_objs_class(
         else:
             raise ValueError(f'{obj_type} type is not supported')
 
-        return process_assembly_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients)
+        return process_assembly_objs(primary_wsid, top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients)
 
     else:
         raise ValueError(f'{obj_type} type is not supported')
@@ -141,12 +145,12 @@ def update_genome_assembly_objs_class(
 
 # process_genome_objs()
 #
-def process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients):
+def process_genome_objs(primary_wsid, top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients):
     objects_created = []
     updated_genome_refs = dict()
     genomeset_query = False
     any_genome_updated = False
-    top_wsid = top_obj['info'][WSID_I]
+    #top_wsid = top_obj['info'][WSID_I]
     
     if len(upas) > 0:
         genomeset_query = True
@@ -177,7 +181,7 @@ def process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_
         std_lineages = get_std_lineages (this_classification, gtdb_ver, this_taxon_id)
 
         # update and save assembly and give genome obj new assembly upa
-        new_assembly_ref = update_and_save_assembly (assembly_obj, std_lineages, clients)
+        new_assembly_ref = update_and_save_assembly (primary_wsid, assembly_obj, std_lineages, clients)
 
         # update genome obj with std_lineages and new assembly obj ref
         genome_obj['data']['assembly_ref'] = new_assembly_ref
@@ -201,10 +205,10 @@ def process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_
             genome_obj['data']['taxonomy'] = classification[assembly_name]
 
         # save updated genome obj
-        this_wsid = genome_obj['info'][WSID_I]
-        if genomeset_query:
-            this_wsid = top_wsid
-        new_ref = save_genome_obj (this_wsid, genome_name, genome_obj['data'], clients)
+        #this_wsid = genome_obj['info'][WSID_I]
+        #if genomeset_query:
+        #    this_wsid = top_wsid
+        new_ref = save_genome_obj (primary_wsid, genome_name, genome_obj['data'], clients)
         updated_genome_refs[original_upa] = new_ref
         desc = 'Taxonomy unchanged, taxon_assignment added GTDB'
         if this_genome_tax_written:
@@ -214,7 +218,7 @@ def process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_
         
     # update refs in genomeset
     if genomeset_query:
-        new_genomeset_ref = update_and_save_genomeset (top_wsid, genomeset_obj, updated_genome_refs, clients)
+        new_genomeset_ref = update_and_save_genomeset (primary_wsid, genomeset_obj, updated_genome_refs, clients)
         desc = 'Taxonomy unchanged, taxon_assignment added GTDB'
         if any_genome_updated:
             desc = 'Taxonomy and taxon_assignment updated with GTDB'
@@ -225,12 +229,12 @@ def process_genome_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_
 
 # process_assembly_objs()
 #
-def process_assembly_objs(top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients):
+def process_assembly_objs(primary_wsid, top_obj, upa, upas, classification, overwrite_tax, gtdb_ver, taxon_assignment_field, clients):
     objects_created = []
     updated_assembly_refs = dict()
     assemblyset_query = False
     any_assembly_updated = False
-    top_wsid = top_obj['info'][WSID_I]
+    #top_wsid = top_obj['info'][WSID_I]
 
     if len(upas) > 0:
         assemblyset_query = True
@@ -261,7 +265,7 @@ def process_assembly_objs(top_obj, upa, upas, classification, overwrite_tax, gtd
         std_lineages = get_std_lineages (this_classification, gtdb_ver, this_taxon_id)
 
         # update and save assembly
-        new_assembly_ref = update_and_save_assembly (assembly_obj, std_lineages, clients)
+        new_assembly_ref = update_and_save_assembly (primary_wsid, assembly_obj, std_lineages, clients)
         updated_assembly_refs[original_upa] = new_assembly_ref
         desc = 'Added GTDB lineage'
         objects_created.append({'ref': new_assembly_ref, 'description': desc})
@@ -269,7 +273,7 @@ def process_assembly_objs(top_obj, upa, upas, classification, overwrite_tax, gtd
         
     # update refs in assemblyset
     if assemblyset_query and any_assembly_updated:
-        new_assemblyset_ref = update_and_save_assemblyset (top_wsid,
+        new_assemblyset_ref = update_and_save_assemblyset (primary_wsid,
                                                            assemblyset_obj,
                                                            updated_assembly_refs,
                                                            clients)
@@ -311,9 +315,13 @@ def get_std_lineages (this_classification, gtdb_ver, this_taxon_id):
 
 # save_genome_obj ()
 #
-def save_genome_obj (this_wsid, genome_name, genome_obj_data, clients):
+def save_genome_obj (primary_wsid, genome_name, genome_obj_data, clients):
+
+    genome_handle_fields = ['genbank_handle_ref', 'gff_handle_ref']
+    genome_obj_data = fix_unowned_shock_handles (genome_obj_data, genome_handle_fields, clients)
+
     updated_obj_info = clients.dfu().save_objects(
-        { 'id': this_wsid,
+        { 'id': primary_wsid,
           'objects': [{ 'type': 'KBaseGenomes.Genome',
                         'name': genome_name,
                         'data': genome_obj_data
@@ -324,12 +332,15 @@ def save_genome_obj (this_wsid, genome_name, genome_obj_data, clients):
 
 # update_and_save_assembly ()
 #
-def update_and_save_assembly (assembly_obj, std_lineages, clients):
+def update_and_save_assembly (primary_wsid, assembly_obj, std_lineages, clients):
     assembly_obj['data']['std_lineages'] = std_lineages
-    this_wsid = assembly_obj['info'][WSID_I]
+    #this_wsid = assembly_obj['info'][WSID_I]
     assembly_name = assembly_obj['info'][NAME_I]
 
-    updated_assembly_obj_info = clients.dfu().save_objects({ 'id': this_wsid,
+    assembly_handle_fields = ['reads_handle_ref', 'fasta_handle_ref']
+    assembly_obj['data'] = fix_unowned_shock_handles (assembly_obj['data'], assembly_handle_fields, clients)
+    
+    updated_assembly_obj_info = clients.dfu().save_objects({ 'id': primary_wsid,
                                                              'objects': [{ 'type': 'KBaseGenomeAnnotations.Assembly',
                                                                            'name': assembly_name,
                                                                            'data': assembly_obj['data']
@@ -340,7 +351,7 @@ def update_and_save_assembly (assembly_obj, std_lineages, clients):
 
 # update_and_save_genomeset ()
 #
-def update_and_save_genomeset (top_wsid, genomeset_obj, updated_genome_refs, clients):
+def update_and_save_genomeset (primary_wsid, genomeset_obj, updated_genome_refs, clients):
     new_genomeset_data = dict()
     genomeset_name = genomeset_obj['info'][NAME_I]
     genomeset_type = genomeset_obj['info'][TYPE_I].split('-')[0]
@@ -381,7 +392,7 @@ def update_and_save_genomeset (top_wsid, genomeset_obj, updated_genome_refs, cli
 
     # save updated genome set obj
     updated_obj_info = clients.dfu().save_objects(
-        { 'id': top_wsid,
+        { 'id': primary_wsid,
           'objects': [{ 'type': 'KBaseSearch.GenomeSet',
                         'name': genomeset_name,
                         'data': new_genomeset_data
@@ -392,7 +403,7 @@ def update_and_save_genomeset (top_wsid, genomeset_obj, updated_genome_refs, cli
 
 # update_and_save_assemblyset ()
 #
-def update_and_save_assemblyset (top_wsid, assemblyset_obj, updated_assembly_refs, clients):
+def update_and_save_assemblyset (primary_wsid, assemblyset_obj, updated_assembly_refs, clients):
     new_assemblyset_data = dict()
     assemblyset_name = assemblyset_obj['info'][NAME_I]
     assemblyset_type = assemblyset_obj['info'][TYPE_I].split('-')[0]
@@ -419,13 +430,29 @@ def update_and_save_assemblyset (top_wsid, assemblyset_obj, updated_assembly_ref
 
     # save updated assembly set obj
     updated_obj_info = clients.dfu().save_objects(
-        { 'id': top_wsid,
+        { 'id': primary_wsid,
           'objects': [{ 'type': 'KBaseSets.AssemblySet',
                         'name': assemblyset_name,
                         'data': new_assemblyset_data
           }]})[0]
     new_ref = upa_from_info(updated_obj_info)
     return new_ref
+
+
+# fix_unowned_shock_handles ()
+#
+def fix_unowned_shock_handles (obj_data, handle_fields, clients):
+
+    for h_field in handle_fields:
+        if obj_data.get(h_field):
+            h_id = obj_data[h_field]
+            s_id = clients.hs().hids_to_handles ([h_id])[0]['id']
+            own_node_output = clients.dfu().own_shock_node({'shock_id': s_id, 'make_handle': 1})
+            new_s_id = own_node_output['shock_id']
+            new_h_id = own_node_output['handle']['hid']
+            obj_data[h_field] = new_h_id
+
+    return obj_data
 
 
 # upa_from_info ()